+++ packages/build_context_holder/example/lib/main.dart
@@ -70,7 +70,7 @@ class _MyHomePageState extends State<MyHomePage>
               style: Theme.of(context).textTheme.headline4,
             ),
             RaisedButton.icon(
-              icon: Icon(Icons.info_outline),
+              icon: const Icon(Icons.info_outline),
               label: const Text('Get information'),
               onPressed: () => _showInfoDialog(_counter),
             ),
@@ -80,7 +80,7 @@ class _MyHomePageState extends State<MyHomePage>
       floatingActionButton: FloatingActionButton(
         onPressed: _incrementCounter,
         tooltip: 'Increment',
-        child: Icon(Icons.add),
+        child: const Icon(Icons.add),
       ),
     );
   }