+++ packages/mwwm/example/lib/ui/main_screen/pages/repositories/repositories_dialog_owner.dart
@@ -1,3 +1,17 @@
+// Copyright (c) 2019-present,  SurfStudio LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import 'package:flutter/cupertino.dart';
 import 'package:flutter/material.dart';
 import 'package:mwwm/mwwm.dart';
@@ -30,9 +44,9 @@ Widget _buildSheet(
 }) {
   return Container(
     height: 100,
-    decoration: BoxDecoration(
+    decoration: const BoxDecoration(
       color: Colors.red,
-      borderRadius: const BorderRadius.only(
+      borderRadius: BorderRadius.only(
         topLeft: Radius.circular(16),
         topRight: Radius.circular(16),
       ),
@@ -40,7 +54,7 @@ Widget _buildSheet(
     child: Center(
       child: Text(
         'Hello ${data.testData}',
-        style: TextStyle(
+        style: const TextStyle(
           color: Colors.white,
           fontSize: 42,
         ),