+++ packages/push_notification/example/lib/notification/messaging_service.dart
@@ -1,5 +1,19 @@
+// Copyright (c) 2019-present,  SurfStudio LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import 'package:firebase_messaging/firebase_messaging.dart';
-import 'package:logger/logger.dart';
+import 'package:push_demo/utils/logger.dart';
 import 'package:push_notification/push_notification.dart';
 
 /// Wrapper over [FirebaseMessaging]
@@ -13,9 +27,7 @@ class MessagingService extends BaseMessagingService {
 
   /// request notification permissions for ios platform
   void requestNotificationPermissions() {
-    _messaging.requestNotificationPermissions(
-      const IosNotificationSettings(),
-    );
+    _messaging.requestNotificationPermissions();
   }
 
   /// no need to call. initialization is called inside the [PushHandler]
@@ -69,7 +81,7 @@ class MessagingService extends BaseMessagingService {
     Map<String, dynamic> message,
     MessageHandlerType handlerType,
   ) async {
-    Logger.d('FIREBASE MESSAGE: $handlerType - $message');
+    logger.d('FIREBASE MESSAGE: $handlerType - $message');
     _handleMessage?.call(message, handlerType);
   }
 }