+++ packages/surf_mwwm/example/lib/ui/screen/counter/di/counter.dart
@@ -13,14 +13,14 @@
 // limitations under the License.
 
 import 'package:flutter/material.dart';
-import 'package:injector/injector.dart';
+import 'package:surf_injector/surf_injector.dart';
 
 /// Component для экрана счетчика
 class CounterComponent implements Component {
-  final NavigatorState navigator;
-  final GlobalKey<ScaffoldState> scaffoldKey = GlobalKey();
-
   CounterComponent(
     this.navigator,
   );
+
+  final NavigatorState navigator;
+  final GlobalKey<ScaffoldState> scaffoldKey = GlobalKey();
 }