+++ packages/swipe_refresh/lib/src/swipe_refresh.dart
@@ -1,3 +1,17 @@
+// Copyright (c) 2019-present,  SurfStudio LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import 'dart:io';
 
 import 'package:flutter/cupertino.dart';
@@ -25,6 +39,7 @@ class SwipeRefresh extends StatelessWidget {
     this.childrenDelegate,
     this.padding,
     this.indicatorColor,
+    this.shrinkWrap = false,
     Color backgroundColor,
     double refreshTriggerPullDistance,
     double refreshIndicatorExtent,
@@ -35,24 +50,25 @@ class SwipeRefresh extends StatelessWidget {
         refreshIndicatorExtent = refreshIndicatorExtent ??
             CupertinoSwipeRefresh.defaultRefreshIndicatorExtent,
         indicatorBuilder = indicatorBuilder ??
-            CupertinoSliverRefreshControl.buildSimpleRefreshIndicator,
+            CupertinoSliverRefreshControl.buildRefreshIndicator,
         super(key: key);
 
   /// Create refresh indicator adaptive to platform.
-  const SwipeRefresh.adaptive(
-      {Key key,
-      List<Widget> children,
-      Stream<SwipeRefreshState> stateStream,
-      SwipeRefreshState initState,
-      VoidCallback onRefresh,
-      Color indicatorColor,
-      Color backgroundColor,
-      double refreshTriggerPullDistance,
-      double refreshIndicatorExtent,
-      RefreshControlIndicatorBuilder indicatorBuilder,
-      ScrollController scrollController,
-      EdgeInsets padding})
-      : this(
+  const SwipeRefresh.adaptive({
+    Key key,
+    List<Widget> children,
+    Stream<SwipeRefreshState> stateStream,
+    SwipeRefreshState initState,
+    VoidCallback onRefresh,
+    Color indicatorColor,
+    Color backgroundColor,
+    double refreshTriggerPullDistance,
+    double refreshIndicatorExtent,
+    RefreshControlIndicatorBuilder indicatorBuilder,
+    ScrollController scrollController,
+    EdgeInsets padding,
+    bool shrinkWrap = false,
+  }) : this(
           SwipeRefreshStyle.adaptive,
           key: key,
           children: children,
@@ -66,6 +82,7 @@ class SwipeRefresh extends StatelessWidget {
           indicatorBuilder: indicatorBuilder,
           scrollController: scrollController,
           padding: padding,
+          shrinkWrap: shrinkWrap,
         );
 
   /// Create refresh indicator with Material Design style.
@@ -79,6 +96,7 @@ class SwipeRefresh extends StatelessWidget {
     Color backgroundColor,
     ScrollController scrollController,
     EdgeInsets padding,
+    bool shrinkWrap = false,
   }) : this(
           SwipeRefreshStyle.material,
           key: key,
@@ -90,6 +108,7 @@ class SwipeRefresh extends StatelessWidget {
           backgroundColor: backgroundColor,
           scrollController: scrollController,
           padding: padding,
+          shrinkWrap: shrinkWrap,
         );
 
   /// Create refresh indicator with Cupertino style.
@@ -104,6 +123,7 @@ class SwipeRefresh extends StatelessWidget {
     RefreshControlIndicatorBuilder indicatorBuilder,
     ScrollController scrollController,
     EdgeInsets padding,
+    bool shrinkWrap = false,
   }) : this(
           SwipeRefreshStyle.cupertino,
           key: key,
@@ -116,6 +136,7 @@ class SwipeRefresh extends StatelessWidget {
           indicatorBuilder: indicatorBuilder,
           scrollController: scrollController,
           padding: padding,
+          shrinkWrap: shrinkWrap,
         );
 
   /// Crete SwipeRefresh as common link
@@ -135,6 +156,7 @@ class SwipeRefresh extends StatelessWidget {
     RefreshControlIndicatorBuilder indicatorBuilder,
     ScrollController scrollController,
     EdgeInsets padding,
+    bool shrinkWrap = false,
   }) {
     assert(itemBuilder != null);
     assert(itemCount != null);
@@ -152,12 +174,10 @@ class SwipeRefresh extends StatelessWidget {
       indicatorBuilder: indicatorBuilder,
       scrollController: scrollController,
       padding: padding,
+      shrinkWrap: shrinkWrap,
       childrenDelegate: SliverChildBuilderDelegate(
         itemBuilder,
         childCount: itemCount,
-        addAutomaticKeepAlives: true,
-        addRepaintBoundaries: true,
-        addSemanticIndexes: true,
       ),
     );
   }
@@ -175,6 +195,7 @@ class SwipeRefresh extends StatelessWidget {
   final ScrollController scrollController;
   final SliverChildDelegate childrenDelegate;
   final EdgeInsets padding;
+  final bool shrinkWrap;
 
   @override
   Widget build(BuildContext context) {
@@ -190,9 +211,11 @@ class SwipeRefresh extends StatelessWidget {
           stateStream: stateStream,
           initState: initState,
           onRefresh: onRefresh,
+          scrollController: scrollController,
           backgroundColor: backgroundColor,
           indicatorColor: indicatorColor,
           children: children,
+          shrinkWrap: shrinkWrap,
         );
       case SwipeRefreshStyle.cupertino:
         return CupertinoSwipeRefresh(
@@ -201,10 +224,12 @@ class SwipeRefresh extends StatelessWidget {
           stateStream: stateStream,
           initState: initState,
           onRefresh: onRefresh,
+          scrollController: scrollController,
           refreshIndicatorExtent: refreshIndicatorExtent,
           refreshTriggerPullDistance: refreshTriggerPullDistance,
           indicatorBuilder: indicatorBuilder,
           children: children,
+          shrinkWrap: shrinkWrap,
         );
       case SwipeRefreshStyle.builder:
       case SwipeRefreshStyle.adaptive: