+++ packages/tabnavigator/lib/src/tab_navigator.dart
@@ -1,3 +1,17 @@
+// Copyright (c) 2019-present,  SurfStudio LLC
+//
+// Licensed under the Apache License, Version 2.0 (the "License");
+// you may not use this file except in compliance with the License.
+// You may obtain a copy of the License at
+//
+//     http://www.apache.org/licenses/LICENSE-2.0
+//
+// Unless required by applicable law or agreed to in writing, software
+// distributed under the License is distributed on an "AS IS" BASIS,
+// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+// See the License for the specific language governing permissions and
+// limitations under the License.
+
 import 'dart:async';
 
 import 'package:flutter/material.dart';
@@ -28,6 +42,7 @@ class TabNavigator extends StatefulWidget {
     this.observersBuilder,
     RouteTransitionsBuilder transitionsBuilder,
     this.transitionDuration = const Duration(milliseconds: 300),
+    this.onGenerateRoute,
   })  : assert(mappedTabs != null),
         assert(selectedTabStream != null),
         assert(initialTab != null),
@@ -41,6 +56,7 @@ class TabNavigator extends StatefulWidget {
   final ObserversBuilder observersBuilder;
   final RouteTransitionsBuilder transitionsBuilder;
   final Duration transitionDuration;
+  final RouteFactory onGenerateRoute;
 
   static TabNavigatorState of(BuildContext context) {
     final Type type = _typeOf<TabNavigatorState>();
@@ -87,6 +103,9 @@ class TabNavigatorState extends State<TabNavigator> {
       stream: widget.selectedTabStream,
       initialData: widget.initialTab,
       builder: (context, snapshot) {
+        if (snapshot.connectionState != ConnectionState.active) {
+          return const SizedBox();
+        }
         final TabType tabType = snapshot.data;
         if (!_initializedTabs.contains(tabType)) {
           _initializedTabs.add(tabType);
@@ -125,22 +144,24 @@ class TabNavigatorState extends State<TabNavigator> {
               observers: widget.observersBuilder != null
                   ? widget.observersBuilder(tabType)
                   : [],
-              onGenerateRoute: (rs) => PageRouteBuilder<Object>(
-                settings: RouteSettings(
-                  name: Navigator.defaultRouteName,
-                ),
-                transitionsBuilder: widget.transitionsBuilder,
-                transitionDuration: widget.transitionDuration,
-                pageBuilder: (
-                  context,
-                  animation,
-                  secondaryAnimation,
-                ) =>
-                    widget.mappedTabs[tabType](),
-              ),
+              onGenerateRoute: (rs) => rs.name == Navigator.defaultRouteName
+                  ? PageRouteBuilder<Object>(
+                      settings: const RouteSettings(
+                        name: Navigator.defaultRouteName,
+                      ),
+                      transitionsBuilder: widget.transitionsBuilder,
+                      transitionDuration: widget.transitionDuration,
+                      pageBuilder: (
+                        context,
+                        animation,
+                        secondaryAnimation,
+                      ) =>
+                          widget.mappedTabs[tabType](),
+                    )
+                  : widget.onGenerateRoute?.call(rs),
             ),
           ),
-        )
+        ),
     ];
   }
 }