+++ tools/ci/lib/tasks/tasks.dart
@@ -103,15 +103,13 @@ Future<void> addCopyrights(
   var filesToCheck = <FileSystemEntity>[];
 
   elements.forEach(
-    (e) async {
-      filesToCheck
-        ..addAll(
-          await fileSystemManager.getEntitiesInModule(
-            e,
-            recursive: true,
-            filter: licenseManager.isNeedCopyright,
-          ),
-        );
+    (element) {
+      List<FileSystemEntity> list = fileSystemManager.getEntitiesInModule(
+        element,
+        recursive: true,
+        filter: licenseManager.isNeedCopyright,
+      );
+      filesToCheck.addAll(list);
     },
   );
 
@@ -148,7 +146,7 @@ Future<void> addCopyrights(
   }
 
   if (troublesList.isNotEmpty) {
-    var errorString;
+    var errorString = '';
 
     troublesList.forEach((key, value) {
       errorString += key.path + ':\n';
@@ -182,13 +180,30 @@ Future<void> mirrorOpenSourceModules(
 ) async {
   final hasRepo = (Element e) => e.openSourceInfo?.separateRepoUrl != null;
   final openSourceModules = elements.where(hasRepo).toList();
+  var failedModulesNames = <String>[];
 
   openSourceModules.forEach(
-    (e) {
-      print('Mirror package ${e.name} to ${e.openSourceInfo.separateRepoUrl}');
-      return MirrorOpenSourceModuleTask(e, currentBranch).run();
+    (element) {
+      print(
+          'Mirror package ${element.name} to ${element.openSourceInfo.separateRepoUrl}');
+      try {
+        return MirrorOpenSourceModuleTask(element, currentBranch).run();
+      } on ModuleMirroringException catch (exception) {
+        print('Package ${element.name} mirroring error !!!');
+        print(exception.message);
+        failedModulesNames.add(element.name);
+      } catch (exception) {
+        print('Package ${element.name} mirroring error !!!');
+        print(exception.toString());
+        failedModulesNames.add(element.name);
+        rethrow;
+      }
     },
   );
+
+  if (failedModulesNames.isNotEmpty) {
+    print('Modules were not mirrored: ${failedModulesNames.join(',')}');
+  }
 }
 
 /// Обновляет зависимости переданных модулей по данным тега.