+++ packages/build_context_holder/CHANGELOG.md
@@ -1,3 +1,11 @@
-## [0.0.1] - TODO: Add release date.
+## 0.0.1-dev.5
 
-* TODO: Describe initial release.
+* Fix lint hints
+
+## 0.0.1-dev.1
+
+* Update README
+
+## 0.0.1-dev.0
+
+* Initial release