+++ packages/push_notification/CHANGELOG.md
@@ -1,3 +1,7 @@
-## 0.0.1
+## 0.0.1-dev.3
 
-* TODO: Describe initial release.
+* Fix lint hints
+
+## 0.0.1-dev.0
+
+* Initial release