+++ packages/relation/CHANGELOG.md
@@ -1,3 +1,12 @@
-## [0.0.2] - TODO: Add release date.
+## 0.0.2-dev.5
 
-* TODO: Describe initial release.
+* Added previous value to loadState
+* Added error to data builder
+
+## 0.0.2-dev.4
+
+* Fix lint hints
+
+## 0.0.2 - 21.0.2020
+
+* Initial release